+++ modules/interrogate/vqa.py
@@ -130,7 +130,7 @@ def qwen(question: str, image: Image.Image, repo: str = None, system_prompt: str
         )
         processor = transformers.AutoProcessor.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         if 'LLM' in shared.opts.cuda_compile:
-            model = compile_torch(model)
+            model = sd_models_compile.compile_torch(model)
         loaded = repo
         devices.torch_gc()
     sd_models.move_model(model, devices.device)
@@ -180,7 +180,7 @@ def gemma(question: str, image: Image.Image, repo: str = None, system_prompt: st
             **quant_args,
         )
         if 'LLM' in shared.opts.cuda_compile:
-            model = compile_torch(model)
+            model = sd_models_compile.compile_torch(model)
         processor = transformers.AutoProcessor.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         loaded = repo
         devices.torch_gc()
@@ -307,7 +307,7 @@ def smol(question: str, image: Image.Image, repo: str = None, system_prompt: str
             )
         processor = transformers.AutoProcessor.from_pretrained(repo, cache_dir=shared.opts.hfcache_dir)
         if 'LLM' in shared.opts.cuda_compile:
-            model = compile_torch(model)
+            model = sd_models_compile.compile_torch(model)
         loaded = repo
         devices.torch_gc()
     sd_models.move_model(model, devices.device)